chore: use venv to install python dependencies (#941)
* chore: use venv to install python dependencies

* chore(scripts): create credentials

* chore: setup exm container

+++ scripts/docker_start.sh
@@ -181,7 +181,9 @@ for container in "${PRE_CONTAINERS[@]}"; do
 done
 
 for container in "${SETUP_CONTAINERS[@]}"; do
-    docker-compose run --rm $container setup
+    if [[ $container != "exm" ]]; then
+        docker-compose run --rm $container setup
+    fi
     start_container $container
 done
